Moved dictionary feature tests out into the integration_tests test suite
